project1/src/requirements.txt
torch>=2.0
numpy
pyyaml
mlflow>=2.0
